parser/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
duckdb
sqlalchemy
pandas
google-genai
python-dotenv
apscheduler
openpyxl
phonenumbers
casparser
rapidfuzz
requests
